html/mail/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
requests
python-dotenv